refactor: [#1211] clean AnnounceHandler tests
